Simplified by removing the overwrites array
Since we are overwriting instead of reapplying all the changes in order, we don't need overwrites array anymore.

Should keep the lock to detect any message coming during the edit. Update the updatedAt for the very rare case that the query shouldn't overwrite everything in the end: a message with later database timestamp.